# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для работы с архивами",
        "title": "Archive API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/archive/documents": {
            "get": {
                "description": "Возвращает список документов с поддержкой пагинации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "archive"
                ],
                "summary": "Получить список документов",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Размер страницы",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestedDocumentsPage"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/controller.ApiError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/controller.ApiError"
                        }
                    }
                }
            }
        },
        "/archive/documents/{id}": {
            "get": {
                "description": "Возвращает документ по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "archive"
                ],
                "summary": "Получить документ по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID документа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ с документом",
                        "schema": {
                            "$ref": "#/definitions/dto.DocumentDto"
                        }
                    },
                    "404": {
                        "description": "Документ не найден",
                        "schema": {
                            "$ref": "#/definitions/controller.ApiError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/controller.ApiError"
                        }
                    }
                }
            },
            "post": {
                "description": "Загружает документ с указанным ID из архивов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "archive"
                ],
                "summary": "Загрузить документ по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID документа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Отправитель запроса",
                        "name": "sender",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Регион",
                        "name": "region",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ c идентификатором созданного запроса",
                        "schema": {
                            "$ref": "#/definitions/dto.CreatedRequest"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/controller.ApiError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/controller.ApiError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "description": {}
            }
        },
        "dto.CreatedRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.DocumentDto": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "document": {},
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.RequestedDocumentsPage": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DocumentDto"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}